Bundle the public LCache library directly
While it's not quite Composer yet, this makes updating the LCache
library much easier.
